double formatting in FieldSolver.cu

+++ FieldSolver.cu
@@ -84,7 +84,7 @@ __global__ void EvaluateFields(const double* dev_potential, double3* dev_el_fiel
 	is_on_low_border = ((threadIdx.x == (blockDim.x - 1)) && (blockIdx.x == (gridDim.x - 1)));
 	is_inside_borders = !(is_on_low_border || is_on_up_border);
 
-	e.x = -((double)1 / ((double)1 + is_inside_borders)) * GradientComponent(
+	e.x = -(1.0 / (1.0 + is_inside_borders)) * GradientComponent(
 		dev_potential[idx + (offset*is_on_up_border) - (offset*is_inside_borders)],
 		dev_potential[idx - (offset*is_on_low_border) + (offset*is_inside_borders)],
 		d_cell_size[0].x);
@@ -94,7 +94,7 @@ __global__ void EvaluateFields(const double* dev_potential, double3* dev_el_fiel
 	is_on_low_border = ((threadIdx.y == (blockDim.y - 1)) && (blockIdx.y == (gridDim.y - 1)));
 	is_inside_borders = !(is_on_low_border || is_on_up_border);
 
-	e.y = -((double)1 / ((double)1 + is_inside_borders)) * GradientComponent(
+	e.y = -(1.0 / (1.0 + is_inside_borders)) * GradientComponent(
 		dev_potential[idx + (offset*is_on_up_border) - (offset*is_inside_borders)],
 		dev_potential[idx - (offset*is_on_low_border) + (offset*is_inside_borders)],
 		d_cell_size[0].y);
@@ -104,7 +104,7 @@ __global__ void EvaluateFields(const double* dev_potential, double3* dev_el_fiel
 	is_on_low_border = ((threadIdx.z == (blockDim.z - 1)) && (blockIdx.z == (gridDim.z - 1)));
 	is_inside_borders = !(is_on_low_border || is_on_up_border);
 
-	e.z = -((double)1 / ((double)1 + is_inside_borders)) * GradientComponent(
+	e.z = -(1.0 / (1.0 + is_inside_borders)) * GradientComponent(
 		dev_potential[idx + (offset * is_on_up_border) - (offset * is_inside_borders)],
 		dev_potential[idx - (offset * is_on_low_border) + (offset * is_inside_borders)],
 		d_cell_size[0].z);
